+++ src/main/java/chess/dao/ChessGameDao.java
@@ -1,6 +1,6 @@
 package chess.dao;
 
-import chess.dao.mapper.PieceMapper;
+import chess.dao.convertor.PiecesConvertor;
 import chess.domain.attribute.Square;
 import chess.domain.chessboard.Chessboard;
 import chess.domain.piece.Piece;
@@ -84,7 +84,7 @@ public final class ChessGameDao {
             PreparedStatement preparedStatement = connection.prepareStatement(query);
             final var resultSet = preparedStatement.executeQuery();
             while (resultSet.next()) {
-                Piece piece = PieceMapper.mapToDomain(new PieceDao(resultSet.getString("piece_type"),
+                Piece piece = PiecesConvertor.mapToDomain(new PieceDao(resultSet.getString("piece_type"),
                         resultSet.getString("piece_team"),
                         resultSet.getString("piece_file"),
                         resultSet.getString("piece_rank")));
