+++ src/main/java/chess/dao/convertor/PiecesConvertor.java
@@ -1,4 +1,4 @@
-package chess.dao.mapper;
+package chess.dao.convertor;
 
 import chess.dao.PieceDao;
 import chess.domain.attribute.Color;
@@ -7,7 +7,7 @@ import chess.domain.attribute.Rank;
 import chess.domain.attribute.Square;
 import chess.domain.piece.Piece;
 
-public class PieceMapper {
+public class PiecesConvertor {
 
     public static Piece mapToDomain(PieceDao pieceDao) {
         Color color = Color.of(pieceDao.team());
